chore: update minimum SDK version to 23 and remove legacy AES_CBC cipher implementation
