+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/ciphers/StorageCipherFactory.java
@@ -20,8 +20,7 @@ enum KeyCipherAlgorithm {
 }
 
 enum StorageCipherAlgorithm {
-    AES_CBC_PKCS7Padding(StorageCipher18Implementation::new, 1),
-    @SuppressWarnings({"UnusedDeclaration"})
+    // Only GCM is supported now (minSdk = 23)
     AES_GCM_NoPadding(StorageCipherGCMImplementation::new, Build.VERSION_CODES.M);
     final StorageCipherFunction storageCipher;
     final int minVersionCode;
@@ -46,37 +45,57 @@ public class StorageCipherFactory {
     private static final String ELEMENT_PREFERENCES_ALGORITHM_PREFIX = "FlutterSecureSAlgorithm";
     private static final String ELEMENT_PREFERENCES_ALGORITHM_KEY = ELEMENT_PREFERENCES_ALGORITHM_PREFIX + "Key";
     private static final String ELEMENT_PREFERENCES_ALGORITHM_STORAGE = ELEMENT_PREFERENCES_ALGORITHM_PREFIX + "Storage";
-    private static final KeyCipherAlgorithm DEFAULT_KEY_ALGORITHM = KeyCipherAlgorithm.RSA_ECB_PKCS1Padding;
-    private static final StorageCipherAlgorithm LEGACY_DEFAULT_STORAGE_ALGORITHM = StorageCipherAlgorithm.AES_CBC_PKCS7Padding;
 
+    // Defaults
+    private static final KeyCipherAlgorithm DEFAULT_KEY_ALGORITHM = KeyCipherAlgorithm.RSA_ECB_PKCS1Padding;
+    private static final StorageCipherAlgorithm DEFAULT_STORAGE_ALGORITHM = StorageCipherAlgorithm.AES_GCM_NoPadding;
+
+    // Note: savedStorageAlgorithm might still read AES_CBC_PKCS7Padding from old prefs, but it won't be used
+    // for new encryption/decryption if requiresReEncryption() is handled correctly.
     private final KeyCipherAlgorithm savedKeyAlgorithm;
     private final StorageCipherAlgorithm savedStorageAlgorithm;
     private final KeyCipherAlgorithm currentKeyAlgorithm;
     private final StorageCipherAlgorithm currentStorageAlgorithm;
 
     public StorageCipherFactory(SharedPreferences source, Map<String, Object> options) {
+        // === Saved Algorithm Determination ===
+        // Read saved algorithms, using GCM as the default if the preference key is missing
+        // Note: This might try to read "AES_CBC_PKCS7Padding" from old prefs, causing an exception if that enum value is removed entirely.
+        // Keeping the enum value but making it unused in selection logic is safer for migration.
+        // If strict cleanup is desired and migration from CBC isn't needed, the AES_CBC entry could be fully removed.
+        String savedStorageAlgoName = source.getString(ELEMENT_PREFERENCES_ALGORITHM_STORAGE, DEFAULT_STORAGE_ALGORITHM.name());
+        try {
+            savedStorageAlgorithm = StorageCipherAlgorithm.valueOf(savedStorageAlgoName);
+        } catch (IllegalArgumentException e) {
+            // Handle case where old prefs had AES_CBC_PKCS7Padding which no longer exists
+            // or if prefs are corrupted. Default to GCM.
+            // Log.w("StorageCipherFactory", "Invalid saved storage algorithm '" + savedStorageAlgoName + "', defaulting to GCM.");
+            savedStorageAlgorithm = DEFAULT_STORAGE_ALGORITHM;
+        }
         savedKeyAlgorithm = KeyCipherAlgorithm.valueOf(source.getString(ELEMENT_PREFERENCES_ALGORITHM_KEY, DEFAULT_KEY_ALGORITHM.name()));
-        savedStorageAlgorithm = StorageCipherAlgorithm.valueOf(source.getString(ELEMENT_PREFERENCES_ALGORITHM_STORAGE, LEGACY_DEFAULT_STORAGE_ALGORITHM.name()));
 
-        StorageCipherAlgorithm targetStorageAlgorithm;
+
+        // === Current Algorithm Determination ===
+
+        // --- Determine and Set Current Storage Algorithm ---
+        // Always use GCM as minSdk is 23+. Check if options specify something else (though only GCM is valid now).
         String storageOptionValue = (String) options.get("storageCipherAlgorithm");
-
+        StorageCipherAlgorithm requestedStorageAlgorithm = DEFAULT_STORAGE_ALGORITHM;
         if (storageOptionValue != null) {
-            targetStorageAlgorithm = StorageCipherAlgorithm.valueOf(storageOptionValue);
-        } else {
-            if (Build.VERSION.SDK_INT >= StorageCipherAlgorithm.AES_GCM_NoPadding.minVersionCode) {
-                targetStorageAlgorithm = StorageCipherAlgorithm.AES_GCM_NoPadding;
-            } else {
-                targetStorageAlgorithm = StorageCipherAlgorithm.AES_CBC_PKCS7Padding;
+            try {
+                requestedStorageAlgorithm = StorageCipherAlgorithm.valueOf(storageOptionValue);
+            } catch (IllegalArgumentException e) {
+                // Log.w("StorageCipherFactory", "Invalid storage algorithm option '" + storageOptionValue + "', using default GCM.");
+                requestedStorageAlgorithm = DEFAULT_STORAGE_ALGORITHM;
             }
         }
+        // Since minSdk=23, requested algorithm's minVersionCode check isn't strictly needed, but good practice.
+        // And currently, only AES_GCM_NoPadding is defined anyway.
+        currentStorageAlgorithm = (requestedStorageAlgorithm.minVersionCode <= Build.VERSION.SDK_INT) ? requestedStorageAlgorithm : DEFAULT_STORAGE_ALGORITHM;
 
-        if (targetStorageAlgorithm.minVersionCode <= Build.VERSION.SDK_INT) {
-            currentStorageAlgorithm = targetStorageAlgorithm;
-        } else {
-            currentStorageAlgorithm = StorageCipherAlgorithm.AES_CBC_PKCS7Padding;
-        }
 
+        // --- Determine and Set Current Key Algorithm ---
+        // Logic remains the same, defaulting to RSA_ECB_PKCS1Padding unless option overrides and is supported.
         final KeyCipherAlgorithm currentKeyAlgorithmTmp = KeyCipherAlgorithm.valueOf(getFromOptionsWithDefault(options, "keyCipherAlgorithm", DEFAULT_KEY_ALGORITHM.name()));
         currentKeyAlgorithm = (currentKeyAlgorithmTmp.minVersionCode <= Build.VERSION.SDK_INT) ? currentKeyAlgorithmTmp : DEFAULT_KEY_ALGORITHM;
     }
@@ -87,11 +106,18 @@ public class StorageCipherFactory {
     }
 
     public boolean requiresReEncryption() {
-        return savedKeyAlgorithm != currentKeyAlgorithm || savedStorageAlgorithm != currentStorageAlgorithm;
+        // Re-encrypt if the key algorithm changed, or if the saved storage algorithm wasn't GCM
+        return savedKeyAlgorithm != currentKeyAlgorithm || savedStorageAlgorithm != StorageCipherAlgorithm.AES_GCM_NoPadding;
     }
 
+    // This method might fail if it tries to instantiate a StorageCipher for an algorithm
+    // that no longer exists (like AES_CBC). Consider if this method is still needed
+    // or how re-encryption handles reading old data.
+    // For now, it assumes valueOf() would fail above if AES_CBC was read and the enum was removed.
     public StorageCipher getSavedStorageCipher(Context context) throws Exception {
         final KeyCipher keyCipher = savedKeyAlgorithm.keyCipher.apply(context);
+        // If savedStorageAlgorithm could be AES_CBC, this line needs careful handling
+        // if that enum value/implementation is removed.
         return savedStorageAlgorithm.storageCipher.apply(context, keyCipher);
     }
 
@@ -102,7 +128,7 @@ public class StorageCipherFactory {
 
     public void storeCurrentAlgorithms(SharedPreferences.Editor editor) {
         editor.putString(ELEMENT_PREFERENCES_ALGORITHM_KEY, currentKeyAlgorithm.name());
-        editor.putString(ELEMENT_PREFERENCES_ALGORITHM_STORAGE, currentStorageAlgorithm.name());
+        editor.putString(ELEMENT_PREFERENCES_ALGORITHM_STORAGE, currentStorageAlgorithm.name()); // Should always be GCM now
     }
 
     public void removeCurrentAlgorithms(SharedPreferences.Editor editor) {
